Create working calculator similar to odin's student example
There are a few bugs, like pressing '=' before entering a complete
expression causing a few problems; as well as pressing any single
operator multiple times causing the same number to undergo the operation
with itself as many times as the operator is pressed. Not sure if
these are really bugs or are normal.

In any case, the functioning of the calculator appears to be identical
to that of the example provided by the odin project.